LayoutTests/accessibility/alt-tag-on-image-with-nonimage-role-expected.txt

   
This tests that setting a role on an img still allows the alt attribute to be used for the description..

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.stringValue is 'AXValue: TEST1'
PASS group.description is 'AXDescription: TEST2'
PASS button.description is 'AXDescription: TEST3'
PASS successfullyParsed is true

TEST COMPLETE
